fixed crash when version jar doesnt exist

+++ scripts/mc.js
@@ -263,10 +263,12 @@ function run(id, software, version, addons, cmd, em, isNew, modpackURL) {
   console.log(software);
 
   if (software != "quilt") {
-    fs.copyFileSync(
-      "data/" + software + "-" + version + ".jar",
-      folder + "/server.jar"
-    );
+    if (fs.existsSync("data/" + software + "-" + version + ".jar")){
+      fs.copyFileSync(
+        "data/" + software + "-" + version + ".jar",
+        folder + "/server.jar"
+      );
+    }
   } else {
     fs.copyFileSync("data/" + software + "-0.5.1.jar", folder + "/server.jar");
     args = [
